feat(helpers): move steps and STATUS_ACTIONS to constantValues
